Cleanup + optimization + minor optimization
Removed unnecessary lea in drawRow I believe it was? May be able to
further optimize in future if we can get to a point where our count is
already in BX and we don't have to copy it from CX.

Minor optimization: commented out "unnecessary" xor BX, BX lines. I
tried the code on real hardware and it seems the BIOS calls don't change
the value in BX. But honestly there's no guarantee this is true for all
IBM compatible PCs, all it tells me is that the Tandy 1000 TL/2 can do
it... hence why they are commented out but not removed, just in case I
end up needing them again.