Valid moves only if destination blank or opponent.
Moves like a1->a2 are no longer valid. The destination must either be
empty, or contain an opponent's piece.

This may need to be changed later to account for more advanced/obscure
chess moves? If so we can just remove it and let the main chess logic do
its job...