Basic knight movement complete (validity).
Surprisingly little code.
Turns out we can use the difference between the destination and source,
and ignore the negative sign (absolute value).
If the (absolute) difference is 3 or greater, the move is invalid (too
far away for any move). Also if the difference is 0 for either
direction, that's invalid, since a knight will always end up on a
different row and column than before.
Then to verify the L shape just make sure the horizontal and vertical
difference are different. If they're the same then the player is trying
to move the piece diagonally!