Move to full 64-byte board. Compression sucks.
Many issues to rant about with the old 32-byte board idea. We might be
on a system with limited RAM, but it won't be *that* limited...

pushing and popping the flags register a million times, shifting nibbles
and masking them too? And all for what, 32 bytes spared? Stupid!

Sure, we could read 4 pieces in at once for drawing the screen, but you
know what that's a sacrifice we're going to have to make - now we can
only read 2 - the horror, the horror!